+++ include/build.h
@@ -2,6 +2,6 @@
  * Author: sascha_lammers@gmx.de
  */
 #pragma once
-#define __BUILD_NUMBER "10401"
-#define __BUILD_NUMBER_INT 10401
+#define __BUILD_NUMBER "10550"
+#define __BUILD_NUMBER_INT 10550