+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.0.6
 
+ - Updated weather station plugin for Arduino Core 3.x
+ - Fixed memory leak in GFXCanvas
  - New auto discovery script ``./scripts/tools/kfcfw_discovery.py``
  - Fixed issues with Home Assistant's name/friendly_name for templates
  - Option to reset factory settings and format file system, if the device keeps crashing and cannot be accessed anymore